requirements/base.txt
asgiref==3.6.0
Django==4.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
python-environ==0.4.54
pytz==2023.3
sqlparse==0.4.3
psycopg2-binary==2.9.5
drf-yasg==1.21.5
django-parler==2.3
Pillow==9.5.0
pre-commit==3.3.2
eskiz-sms==0.2.3
#filters
django-filter==23.2

# testing
factory-boy==3.2.1
model_bakery==1.11.0
pytest==7.3.1
pytest-mock==3.11.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
pytest-env==0.8.2
# caches
django-redis==5.2.0
redis==4.5.3

#channels
daphne==4.0.0
channels==4.0.0
channels_redis==4.1.0